requirements.txt
flask-shell-ipython==0.4.1
flask-sqlalchemy==2.4.0
flask-wtf==0.14.2
ipdb==0.12.2
pip-chill==0.1.8
